### 2024-05-03 #6718 @editorialbot
[PRE REVIEW]: DSSE: An environment for simulation of reinforcement learning-empowered drone swarm maritime search and rescue missions
**Submitting author:** @renatex333 (<a href="http://orcid.org/0009-0001-5943-0481">Renato Laffranchi Falcão</a>)
**Repository:** https://github.com/pfeinsper/drone-swarm-search
**Branch with paper.md** (empty if default branch): main
**Version:** 1.1.7
