+++ .github/workflows/draft-pdf.yml
@@ -7,6 +7,7 @@ on:
       - 'main'
     paths:
       - 'paper.md'
+      - 'paper.bib'
 
 jobs:
   paper:
