### 2024-05-03 0fbc825
Fix: prob_matrix_path error
