+++ README.md
@@ -7,7 +7,9 @@
 
 # <img src="https://github.com/pfeinsper/drone-swarm-search/blob/main/docs/public/pics/drone.svg" alt="DSSE Icon" width="45" height="25"> Drone Swarm Search Environment (DSSE)
 
-Welcome to the official GitHub repository for the Drone Swarm Search Environment (DSSE). This project is dedicated to providing a comprehensive simulation environment designed to develop, test, and refine drone swarm search strategies. Here, researchers and developers can access a versatile platform that supports a wide range of drone swarm simulations, enabling the exploration of complex behaviors and interactions within dynamic, real-world scenarios.
+Welcome to the official GitHub repository for the Drone Swarm Search Environment (DSSE). This project offers a comprehensive simulation platform designed for developing, testing, and refining search strategies using drone swarms. Researchers and developers will find a versatile toolset supporting a broad spectrum of simulations, which facilitates the exploration of complex drone behaviors and interactions in dynamic, real-world scenarios.
+
+In this repository, we have implemented two distinct types of environments. The first is a dynamic environment that simulates maritime search and rescue operations for shipwreck survivors. It models the movement of individuals in the sea using a dynamic probability matrix, with the objective for drones being to locate and identify these individuals. The second is a static environment utilizing the Lagrangian particle simulation from the open-source [Opendrift library](https://github.com/OpenDrift/opendrift), which incorporates real-world ocean data to create a probability matrix for drone navigation tasks. In this scenario, drones are tasked with efficiently identifying all relevant cells that meet a certain probability threshold.
 
 
 ## 📚 Documentation Links
@@ -18,20 +20,14 @@ Welcome to the official GitHub repository for the Drone Swarm Search Environment
 
 - **[PyPI Repository](https://pypi.org/project/DSSE/)**: Visit the PyPI page for DSSE to download the latest release, view release histories, and read additional installation instructions.
 
+# DSSE - Search Environment
+
 ## 🎥 Visual Demonstrations
 <p align="center">
-    <img src="https://raw.github.com/PFE-Embraer/drone-swarm-search/env-cleanup/docs/gifs/render_with_grid_gradient.gif" width="400" height="400" align="center">
+    <img src="docs/public/gifs/render_with_grid_gradient.gif" width="400" height="400" align="center">
     <br>
     <em>Above: A simulation showing how drones adjust their search pattern over a grid.</em>
 </p>
-
-## ⚡Quick Start
-
-### ⚙️ Installation
-Install DSSE quickly with pip:
-```bash
-pip install DSSE
-````
 
 ## 🎯 Outcome
 
@@ -39,8 +35,16 @@ pip install DSSE
 :-------------------------:|:-------------------------:
 | ![](https://raw.githubusercontent.com/PFE-Embraer/drone-swarm-search/main/docs/public/pics/victory_render.png)     | ![](https://raw.github.com/PFE-Embraer/drone-swarm-search/main/docs/public/pics/fail_render.png) |
 
+## ⚡ Quick Start
 
-## 🛠️ Basic Env Usage
+### ⚙️ Installation
+Quickly install DSSE using pip:
+```bash
+pip install DSSE
+````
+
+
+## 🛠️ Basic Env Search Usage
 ```python
 from DSSE import DroneSwarmSearch
 
@@ -81,6 +85,25 @@ while not done:
     observations, rewards, terminations, truncations, infos = env.step(actions)
     done = any(terminations.values()) or any(truncations.values())
 ```
+
+
+# DSSE - Coverage Environment
+
+## 🎥 Visual Demonstrations
+<p align="center">
[…]
