### 2024-05-07 0be45b6
Fix docs problems + add packages to install GDAL
