+++ DSSE/environment/env.py
@@ -66,7 +66,7 @@ class DroneSwarmSearch(DroneSwarmSearchBase):
         print(f"Pre render time: {pre_render_time} minutes")
         print(f"Pre render steps: {self.pre_render_steps}")
 
-        # Prob m"DSSE/tests/matrix.npy"atrix
+        # Prob matrix
         self.probability_matrix = None
         self.dispersion_inc = dispersion_inc
         self.dispersion_start = dispersion_start
