### 2024-05-08 1b49805
Modifying reward function on coverage env + noise filter on coverage matrix creation
