### 2024-05-09 41bb533
Adjust opendrift gif
