### 2024-05-09 a1b3097
Separa a busca de pessoas em funções
