+++ DSSE/environment/simulation/particle_sim.py
@@ -1,42 +1,196 @@
-from opendrift.models.oceandrift import OceanDrift
+import math
+import numpy as np
+from datetime import datetime, timedelta
+from typing import List, Tuple
 
 
-def open_drift(lat, lon, time, number, radius, duration, outfile):
-    o = OceanDrift()
-    o.add_readers_from_list(
-        ["https://tds.hycom.org/thredds/dodsC/GLBy0.08/expt_93.0/uv3z"]
-    )
-    o.seed_elements(lat=lat, lon=lon, time=time, number=number, radius=radius)
+EARTH_MEAN_RADIUS = 6373.0
 
-    o.run(duration=duration, outfile=outfile)
 
-    lat_final = o.elements.lat
-    lon_final = o.elements.lon
+class ParticleSimulation:
+    def __init__(
+        self,
+        disaster_lat: float,
+        disaster_long: float,
+        duration_hours: int = 10,
+        loglevel: int = 20,
+        animate: bool = False,
+        cell_size: int = 130,
+    ) -> None:
+        try:
+            from opendrift.models.oceandrift import OceanDrift
+            self.ocean_drift = OceanDrift
+        except ImportError:
+            raise ImportError("OpenDrift not installed. Install the environment with the 'coverage' extra: pip install DSSE[coverage]")
+        
+        self.disaster_lat = disaster_lat
+        self.disaster_long = disaster_long
+        self.loglevel = loglevel
+        self.animate = animate
+        self.duration_hours = duration_hours
+        self.cell_size = cell_size
 
-    return list(zip(lat_final, lon_final))
+        # Internal variables
+        self.map_size = 0
+        self.original_map = None
+        self.probability_map = None
 
-def calculate_bounding_rectangle(coordinates):
-    min_lat = max_lat = coordinates[0][0]
-    min_lon = max_lon = coordinates[0][1]
+    def run_or_get_simulation(self):
+        if self.probability_map is None:
+            self.run_simulation()
+        self.probability_map = self.original_map.copy()
 
-    for lat, lon in coordinates:
-        min_lat = min(min_lat, lat)
-        max_lat = max(max_lat, lat)
-        min_lon = min(min_lon, lon)
-        max_lon = max(max_lon, lon)
+    def run_simulation(self):
+        duration = timedelta(hours=self.duration_hours)
+        start_time = datetime.now() - duration
+        number = 50_000
+        radius = 1000
 
-    return min_lat, max_lat, min_lon, max_lon
+        coordinates = self.simulate(start_time, number, radius, duration)
+        self.map_size = self.calculate_map_size(coordinates)
+        cartesian = self.convert_lat_lon_to_xy(coordinates)
+        self.probability_map = self.create_probability_map(cartesian)
+        # Maintain always a copy of the original map
+        self.original_map = self.probability_map.copy()
 
-def convert_lat_lon_to_xy(coordinates, map_size):
-    min_lat, max_lat, min_lon, max_lon = calculate_bounding_rectangle(coordinates)
-    width = max_lon - min_lon
-    height = max_lat - min_lat
-    map_width, map_height = map_size
+    def simulate(
+        self,
+        time: datetime,
+        number: int,
+        radius: int,
+        duration: timedelta,
+    ) -> List[Tuple[float, float]]:
+        o = self.ocean_drift(loglevel=self.loglevel)
+        o.add_readers_from_list(
+            ["https://tds.hycom.org/thredds/dodsC/GLBy0.08/expt_93.0/uv3z"]
+        )
+        o.seed_elements(
+            lat=self.disaster_lat,
+            lon=self.disaster_long,
+            time=time,
+            number=number,
[…]
