### 2024-05-10 3f12a26
Changing the reward function cov
