### 2024-05-14 b745c2b
negative reward on default action
