### 2024-05-14 #6746 @renatex333
Hi!

I am thrilled to have our paper reviewed by you. I truly appreciate the time and effort you are dedicating to this process.

I wanted to inform you that I have a newer version of the manuscript in which I corrected an error in the bibliography that was present in the originally submitted version. I kindly request your guidance on how I can submit this updated version for your consideration.

If there is anything else I can do to assist in this process or provide further information, please do not hesitate to let me know.

Thank you once again for your valuable time and feedback!
