### 2024-05-15 ea0842a
Fine tuning in the reward function
