+++ requirements.txt
@@ -2,7 +2,7 @@ numpy>=1.23.1
 gymnasium>=0.27.1
 pygame>=2.3.0
 pettingzoo>=1.22.3
-matplotlib>=3.7.0
+matplotlib==3.8.4
 pytest>=8.0.0
 numba>=0.59.0
 wheel
