+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = [
     "gymnasium>=0.27.1",
     "pygame>=2.3.0",
     "pettingzoo>=1.22.3",
-    "matplotlib>=3.7.0",
+    "matplotlib==3.8.4",
     "numba>=0.59.0",
 ]
 
