### 2024-05-21 f0897d7
Configurable cell size for the search env as well
