### 2024-05-21 b11027a
Fix matrix numpy type
Co-authored-by: Renato Laffranchi Falcão <66042074+renatex333@users.noreply.github.com>
