### 2024-05-23 f900e53
New reward and terminations for both envs
