+++ docs/Documentation/docsCoverage.md
@@ -91,18 +91,23 @@ We incorporated 8 actions in this environment to enable the use of agents traine
 :::
 
 ### Inputs
-| Inputs                          | Possible Values       | Default Values            |
-| -------------                   | -------------         | -------------             |
-| `render_mode`                   | `"ansi" or "human"`   | `"ansi"`                  |
-| `render_grid`                   | `bool`                | `True`                    |
-| `render_gradient`               | `bool`                | `True`                    |
-| `timestep_limit`                | `int`                 | `100`                     |
-| `disaster_position`             | `(float, float)`      | `(-24.04, -46.17)`        |
-| `drone_amount`                  | `int`                 | `1`                       |
-| `drone_speed`                   | `int`                 | `10`                      |
-| `drone_probability_of_detection`| `float`               | `1.0`                     |
-| `pre_render_time`               | `int`                 | `10`                      |
-| `prob_matrix_path`              | `string`              | `None`                    |
+| Inputs                            | Possible Values       | Default Values            |
+| -------------                     | -------------         | -------------             |
+| `render_mode`                     | `"ansi" or "human"`   | `"ansi"`                  |
+| `render_grid`                     | `bool`                | `True`                    |
+| `render_gradient`                 | `bool`                | `True`                    |
+| `timestep_limit`                  | `int`                 | `100`                     |
+| `disaster_position`               | `(float, float)`      | `(-24.04, -46.17)`        |
+| `drone_amount`                    | `int`                 | `1`                       |
+| `drone_speed`                     | `int`                 | `10`                      |
+| `drone_probability_of_detection`  | `float`               | `1.0`                     |
+| `pre_render_time`                 | `int`                 | `10`                      |
+| `prob_matrix_path`                | `string`              | `None`                    |
+| `particle_amount`                 | `int`                 | `50,000`                  |
+| `particle_radius`                 | `int`                 | `800`                     |
+| `num_particle_to_filter_as_noise` | `int`                 | `1`                       |
+| `start_time`                      | `datetime`            | `None`                    |
+| `grid_cell_size`                  | `int`                 | `130`                     |
 
 - **`render_mode`**:
 
@@ -126,6 +131,16 @@ We incorporated 8 actions in this environment to enable the use of agents traine
 - **`pre_render_time`**: This **int** parameter specifies the amount of time `(hours)` to pre-render the simulation before starting. Adjusting this value lets the user control the pre-rendering time of the simulation.
 
 - **`prob_matrix_path`**: This **string** parameter allows the user to specify the path to file of a already simulated probability matrix. The file should be a `.npy` file containing a probability matrix. If this parameter is not specified, the environment will generate a new probability matrix.
+
+- **`particle_amount`**: This **int** parameter allows the user to customize the number of particles used in the `Lagrangian particle model` used to create the probability matrix.
+
+- **`particle_radius`**: This **int** parameter allows the user to `customize` the radius that the particles are randomly placed at the start of the Lagrangian particle model simulation.
+
+- **`num_particle_to_filter_as_noise`**: This **int** parameter allows the user to modify the number of particles in each cell that are filtered to zero.  (e.g. if the number of value is `1`, cells with only 1 particle in the end of simulation will be filtered to 0).
+
+- **`start_time`**: This **datetime** parameter allows the user to specify the `start time` of the simulation. If not specified, the simulation will start at the current time.
+
[…]
