### 2024-05-27 fe08939
Update drone svg path to raw
