### 2024-05-28 1f00bf3
algorithms docs
