+++ docs/Documentation/docsAlgorithms.md
@@ -6,27 +6,327 @@ Welcome to our [Algorithms repository](https://github.com/pfeinsper/drone-swarm-
 
 Explore a diverse range of implementations that leverage the latest advancements in machine learning to solve complex coordination tasks in dynamic and unpredictable environments. Our repository offers state-of-the-art solutions designed to enhance the performance and adaptability of autonomous drone swarms, making them more efficient and effective in various search and rescue missions.
 
+| Search Environment       |
+:-------------------------:|
+| ![](/gifs/render_with_grid_gradient.gif)|
+
+| Coverage Environment   |
+:-------------------------:
+![](/gifs/basic_coverage.gif) |
+
+## Algorithms Implemented
+
+In this project, we have utilized the [RLlib library](https://docs.ray.io/en/latest/rllib/index.html), a library designed for reinforcement learning (RL), to develop our algorithms. Our implementations are based on a Multi-Input Convolutional Neural Network (CNN) architecture, adeptly handling the complex dynamics of the environments we are exploring.
+
+### Multi-Input CNN Architecture
+
+The neural network architecture employed here is specifically designed to process multiple types of input data simultaneously, reflecting the comprehensive state of the environment:
+   
+1. **`Probability Matrix`**: This input encapsulates a matrix of probabilities, providing a possible position for the target within the environment.
+   
+2. **`Agent Positions`**: The second input consists of the spatial positions of agents within the environment. This allows the network to integrate spatial awareness with probabilistic insights, offering a nuanced understanding of the environment.
+
+By combining these data, our CNN can effectively interpret and react to the full scope of environmental observations.
+
+### Proximal Policy Optimization (PPO): An Overview
+
+Proximal Policy Optimization (PPO) is a widely used algorithm in the field of reinforcement learning (RL), known for its effectiveness and relative simplicity. It belongs to a class of algorithms called policy gradient methods, which optimize the policy directly. PPO aims to improve training stability and efficiency by addressing the problem of large policy updates, which can destabilize learning in RL.
+
+#### Centralized vs. Decentralized Implementations
+
+PPO can be implemented in two distinct ways: `centralized` and `decentralized`.
+
+1. **`Centralized PPO`**:
+   - In a centralized setup, a single neural network is used for `all agents` in the environment. This network learns a policy that collectively governs the actions of `all agents`. The primary advantage of this approach is that it allows for a unified update mechanism where `all agents` contribute to and benefit from a single learning model, promoting consistency in the learned policy.
+   - Centralized PPO is particularly useful in scenarios where agents’ actions are interdependent, and the state and reward information from all agents can be aggregated to make more informed policy updates.
+
+2. **`Decentralized PPO`**:
+   - Conversely, decentralized PPO employs a **separate neural network** for `each agent`. Each network learns an **individual** policy based on the agent's own experiences. This approach is advantageous in environments where agents have distinct roles or when their observations and objectives differ significantly.
+   - Decentralized PPO supports greater customization of learning and adaptation for `each agent`, potentially leading to more robust performance in diverse or non-uniform settings.
+
+#### Core Mechanism
+
+Regardless of the implementation style, the core mechanism of PPO revolves around optimizing the policy in a way that avoids too large updates. PPO introduces a clipped surrogate objective function, which helps in managing the size of policy updates. This clipping mechanism limits the ratio of new policy probabilities to old policy probabilities, ensuring that the updates are neither too small (leading to slow learning) nor too large (causing instability).
+
+### Deep Q-Network (DQN): Implementation Overview
+
[…]
