+++ docs/OurStory/ourStory.md
@@ -1,17 +1,157 @@
-# Our Story Coming Soon!
+# The Beginning of Our Story
 
-![Under Construction](/pics/build.png)
+[Insper](https://www.insper.edu.br/), a distinguished university located in São Paulo, Brazil, offers an engineering program that includes a comprehensive final project known as "PFE" (Projeto Final de Engenharia). This project, integral to the 8th semester of the engineering course, involves students working in groups of three or four. They collaborate directly with companies to develop their projects over a semester and present their findings to a jury at the semester's end.
 
-We are currently working hard on developing a captivating page that will tell the story of how our project began. This narrative will delve into the origins of the project, the initial ideas, the challenges we faced, and how it has evolved into what it is today.
+A key partnership is between Insper and Embraer, facilitated through the Mei hub platform which connects companies like Embraer with universities to collaborate on developmental projects. Embraer, a major aerospace manufacturer, has chosen to work with Insper on the "Drone Swarm Search" project. This initiative focuses on deploying a swarm of drones to locate a target within a designated area quickly and efficiently.
 
-🛠️ **What to Expect?**
-- **Origins and Evolution**: An engaging story of our project's inception and its transformative journey.
-- **Key Milestones**: Highlights of major achievements and turning points in our project history.
-- **Behind the Scenes**: Insights into the creative and technical processes that drive our project.
+Representing Embraer in this collaboration is Dr. José Fernando Basso Brancalion, while Insper's faculty representative is Dr. Fabrício Barth.
 
-📅 **Estimated Availability**
-- The full story is expected to be available soon. We are committed to making it inspiring, informative, and reflective of our journey.
+<div align="center">
+    <table>
+        <tr>
+            <td align="center">
+                <img src="https://github.com/fbarth.png" width="100px" alt="" style="border-radius: 50%;">
+                <br>
+                <sub style="font-size: 16px; font-weight: none;"><b>Dr. Fabrício Barth</b></sub>
+                <br>
+                <sub style="font-size: 15px; font-weight: bold;">Orientador</sub>
+                <br>
+                <button onclick="window.location.href='https://github.com/fbarth';" style="background: none; border: none; cursor: pointer; padding: 0; margin-right: 10px;">
+                    <img src="https://icons.iconarchive.com/icons/papirus-team/papirus-apps/256/github-icon.png" alt="GitHub" width="30">
+                </button>
+                <button onclick="window.location.href='https://www.linkedin.com/in/fbarth/';" style="background: none; border: none; cursor: pointer; padding: 0;">
+                    <img src="https://icons.iconarchive.com/icons/limav/flat-gradient-social/256/Linkedin-icon.png" alt="LinkedIn" width="25">
+                </button>
+            </td>
+            <td align="center">
+                <img src="https://media.licdn.com/dms/image/C4E03AQGn6DuuYwhjBw/profile-displayphoto-shrink_200_200/0/1564709106422?e=1718236800&v=beta&t=3JrDTb5QTF4k5qFZbQc3lK9sgSJbalH7Y3QD_rthXBE" width="100px" alt="" style="border-radius: 50%;">
+                <br>
+                <sub style="font-size: 16px; font-weight: none;"><b>Dr. José F. B. Brancalion</b></sub>
+                <br>
+                <sub style="font-size: 15px; font-weight: bold;">Representante da Embraer</sub>
+                <br>
+                <button onclick="window.location.href='https://www.linkedin.com/in/jose-fernando-basso-brancalion/';" style="background: none; border: none; cursor: pointer; padding: 0;">
[…]
