+++ docs/Documentation/docsAlgorithms.md
@@ -272,13 +272,13 @@ The following table outlines the hyperparameters used for the Proximal Policy Op
 
 **Table A1: Hyperparameters for PPO**
 
-| Parameter | Value   | Description                           |
-|-----------|---------|---------------------------------------|
-| B         | 8192    | Training batch size                   |
-| Lr        | 10^-5   | Learning rate                         |
-| 𝛾         | 0.9999999 | Discount factor                       |
-| M         | 300     | Stochastic Gradient Descent (SGD) minibatch |
-| K         | 10      | Number of SGD iterations              |
+| Parameter | Value     | Description                                 |
+|-----------|-----------|---------------------------------------------|
+| B         | 8192      | Training batch size                         |
+| Lr        | 10^-5     | Learning rate                               |
+| 𝛾         | 0.9999999 | Discount factor                             |
+| M         | 300       | Stochastic Gradient Descent (SGD) minibatch |
+| K         | 10        | Number of SGD iterations                    |
 
 ### Hyperparameters for DQN
 
@@ -286,15 +286,15 @@ The following table outlines the hyperparameters used for the Deep Q-Network (DQ
 
 **Table A2: Hyperparameters for DQN**
 
-| Parameter | Value    | Description                                   |
-|-----------|----------|-----------------------------------------------|
-| B         | 8192     | Training batch size                           |
-| Lr        | 10^-5    | Learning rate                                 |
-| 𝛾         | 0.9999999 | Discount factor                               |
-| U         | 500      | Update target network every U steps           |
-| 𝜀0        | 1        | Initial epsilon for 𝜀-greedy                  |
-| 𝜀f        | 0.1      | Final epsilon for 𝜀-greedy                    |
-| T         | 400000   | T timesteps for epsilon to decay from 𝜀0 to 𝜀f |
+| Parameter | Value     | Description                                    |
+|-----------|-----------|------------------------------------------------|
+| B         | 8192      | Training batch size                            |
+| Lr        | 10^-5     | Learning rate                                  |
+| 𝛾         | 0.9999999 | Discount factor                                |
+| U         | 500       | Update target network every U steps            |
+| 𝜀0        | 1         | Initial epsilon for 𝜀-greedy                   |
+| 𝜀f        | 0.1       | Final epsilon for 𝜀-greedy                     |
+| T         | 400000    | T timesteps for epsilon to decay from 𝜀0 to 𝜀f |
 
 ### Environment Settings for Search Tests
 
@@ -302,13 +302,13 @@ The following table provides the settings for the search environment used in the
 
 **Table A3: Environment Settings for Search Tests**
 
-| Parameter              | Value          | Description                          |
-|------------------------|----------------|--------------------------------------|
-| Incremento de dispersão | 0.1 (0.05 H1) | Dispersion increment                 |
-| Número de PIW         | 1 (4 H5)       | Number of Persons in Water (PIW)     |
-| Tamanho do grid       | 40x40          | Grid size                            |
-| Número de drones      | 4              | Number of drones                     |
-| Número máximo de timesteps por simulação | 100 | Maximum number of timesteps per simulation |
+| Parameter                                  | Value          |
+|--------------------------------------------|----------------|
+| Dispersion increment                       | 0.1 (0.05 H1)  |
+| Number of Persons in Water (PIW)           | 1 (4 H5)       |
+| Grid size                                  | 40x40          |
+| Number of drones                           | 4              |
[…]
