### 2024-05-29 pfeinsper/drone-swarm-search#258 @draabe
JOSS Review: Contribution and Authorship
Hi!

Regarding this checkbox of the JOSS Review: Contribution and authorship: Has the submitting author (@renatex333) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I tried to match the paper authors with the repository contributors, but couldn't quite resolve the real and user names.

Could you provide me with a brief rundown of all the author contributions?

Thanks!
