### 2024-05-29 pfeinsper/drone-swarm-search#259 @draabe
JOSS Review: package version
Hi,

just a quick question: the package version stated for the JOSS review (https://github.com/openjournals/joss-reviews/issues/6746) is 1.1.7, which is quite off from the last version release (0.2.2).

Am I missing something here?

Thanks!
