### 2024-05-29 pfeinsper/drone-swarm-search#260 @draabe
JOSS Review: paper feedback
Hi,

here are my revision points for the software paper.

**Summary**

I think that, overall, the paper looks really great. It is well-written, concise, and easy-to-follow but instructive even for a non-domain-informed audience. There are a few points (see below) that could use some slight improvements/additions, but I feel the paper is close to checking the JOSS boxes.

**Major points**

- Summary: I think the summary should be extended slightly to discuss a bit more the high-level functionality of your package. The purpose becomes clear, but what exactly can I do as a user, i.e., what are the main functions? What environments can I create, what stochastic simulations can I run, do you provide policies for re-inforcement learning or do I have to provide my own? Ideally, I'd know if your package will help and fit within my research project or product after reading your Summary section. You could also think about summarizing the functionality in a dedicated Functionality section.
- State of the field: Although I acknowledge that you mention other common approaches to SAR planning (particularly path planning), I still feel the state of the field could be described a bit more verbosely. What are the limitations of previous approaches, and how do approaches based on reinforcement learning address these limitations or improve the finding of search strategies? How is a SAR path planning problem typically defined, and how (i.e., with which metrics) is a solution evaluated?
- State of the field (second point): Could you also briefly discuss other solutions (even if not publicly available or proprietary) as far as possible, and compare your package to them (e.g., in terms of functionality and availability).
- One last note on the structure of the paper: Anything after the second paragraph of the Statement of Need section (line 28 and below) does not, in my opinion, really contribute further to the statement of need (which is fully and convincingly stated in the first two paragraphs). So it could be converted into a separate section (something like Package Description or else) for clarity. But I believe this is more a matter of personal preference, so I'll leave the decision to you.

**Minor**

10: the second sentence of the summary is quite long and confusing. maybe split into multiple sentences for simplification.
11: equip's -> equips
25: researches propose -> research proposes
28: researches -> research

**Figures**

- There seems to be an issue with the compilation of Figure 1

**References**

- the link for the reference "International aeronautical and maritime search and rescue manual - volume II - mission84
co-ordination: Vol. II (9th ed.). (2022). (9th ed.) " does not compile nicely, but I'm also not sure how to fix it. Maybe the manual has an ISBN instead?
- Terry, Black, ... (2021) is missing a DOI or URL
