### 2024-05-30 pfeinsper/drone-swarm-search#259 @RicardoRibeiroRodrigues
Hi,
the 1.1.7 version on pypi was the latest version at the time we were submitting the paper, but now we are at 1.1.9 (just some bug fixes em documentation updates). About the releases, we end up forgetting to update the github releases, only made the tags. I just made the github 1.1.9 release, so feel free to use it for review purposes.
Thanks for the feedback.
