### 2024-06-02 pfeinsper/drone-swarm-search#258 @renatex333
Hi,

Sure, I'd be happy to clarify the contributions of each author. Here is a brief rundown:

- @renatex333 (Renato Laffranchi Falcão): Responsible for code refactoring and key improvements in the PIW (Person in Water) movement, enhancing the simulation's realism and overall performance. Contributed to making the project more accessible to new contributors and promoting open-source standards. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @Pedro2712 (Pedro Henrique Britto Aragão Andrade): Responsible for the development of automated tests and the coverage environment. Led efforts in code refactoring and enhancing the simulation environment, including the implementation of multiple shipwrecked people and the Probability of Detection feature. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @RicardoRibeiroRodrigues (Ricardo Ribeiro Rodrigues): Developed the coverage environment, studied ways to implement communication between drones to share verified areas, resolved performance issues with numpy, improved the probability matrix calculation performance, adapted the environment interface to the PettingZoo standard, refactored code, implemented new functionalities in the simulation, restructured code, and implemented Reinforcement Learning algorithms. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @JorasOliveira (Jorás Custódio Campos de Oliveira): Responsible for restructuring and organizing the repository, reviewing reward functions, and refactoring code. Ensured the simulation continues running when a drone attempts to exit the boundary, added a time parameter to track the elapsed time and simulate the individuals positions before the search begins (pre-simulation), added the project to PettingZoo's list of Third Party projects, and made the library available on PyPI. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @Manuel-castanares (Manuel Castanares): Responsible for creating the probability matrix and integrating it with the environment. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @enricofd (Enrico Damiani): Responsible for the development of automated tests, code refactoring, integration of the probability matrix with the environment map and with Pygame functionalities, and implemented Reinforcement Learning algortihms. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @lfcarrete (Luis Filipe Carrete): Responsible for the development of the visualization of the simulation environment with Pygame. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @leonardodma (Leonardo Malta de Abreu): Responsible for implementing Reinforcement Learning algorithms and improving the reward function of the environment to better promote the drones exploration. Contributed significantly to the experiments with Reinforcement Learning algorithms, documentation, papers and reports.
- @fbarth (Fabrício Jailson Barth): Reinforcement Learning researcher and professor who provided essential guidance and mentorship throughout the project's development. Contributed significantly to the experiments with Reinforcement Learning algorithms, as well as to the documentation, papers, and reports.
- José Fernando Basso Brancalion: Provided crucial guidance and mentorship during the project's development. Played a significant role in the experiments with Reinforcement Learning algorithms.

This should help resolve the contributions and ensure that the list of authors is appropriate and complete.

Thanks!
