+++ paper.bib
@@ -18,13 +18,15 @@
   note         = {Accessed: 2024-04-24}
 }
 
-@manual{iamsar,
-  title        = {International Aeronautical and Maritime Search and Rescue Manual - Volume II - Mission Co-ordination},
+@inbook{iamsar,
+  title        = {International Aeronautical and Maritime Search and Rescue Manual - },
   organization = {International Maritime Organization and International Civil Aviation Organization},
+  publisher    = {International Maritime Organization},
   year         = {2022},
   volume       = {II},
   edition      = {9th},
   chapter      = {5},
+  isbn         = {9789280117356},
   url          = {https://store.icao.int/en/international-aeronautical-and-maritime-search-and-rescue-manual-volume-ii-mission-co-ordination-doc-9731-2}
 }
 
@@ -37,7 +39,7 @@
   year    = {2021}
 }
 
-@software{Terry_PettingZoo_Gym_for,
+@misc{Terry_PettingZoo_Gym_for,
   author  = {Terry, Jordan and Black, Benjamin and Grammel, Nathaniel and Jayakumar, Mario and Hari, Ananth and Sullivan, Ryan and Santos, Luis and Perez, Rodrigo and Horsch, Caroline and Dieffendahl, Clemens and Williams, Niall and Lokesh, Yashas},
   license = {MIT},
   title   = {{PettingZoo: Gym for multi-agent reinforcement learning}},
