+++ paper.bib
@@ -19,27 +19,29 @@
 }
 
 @inbook{iamsar,
-  title        = {International Aeronautical and Maritime Search and Rescue Manual - },
+  title        = {International Aeronautical and Maritime Search and Rescue Manual},
   organization = {International Maritime Organization and International Civil Aviation Organization},
   publisher    = {International Maritime Organization},
   year         = {2022},
   volume       = {II},
-  edition      = {9th},
   chapter      = {5},
   isbn         = {9789280117356},
   url          = {https://store.icao.int/en/international-aeronautical-and-maritime-search-and-rescue-manual-volume-ii-mission-co-ordination-doc-9731-2}
 }
 
-@article{terry2021pettingzoo,
-  title   = {Pettingzoo: Gym for multi-agent reinforcement learning},
-  author  = {Terry, J and Black, Benjamin and Grammel, Nathaniel and Jayakumar, Mario and Hari, Ananth and Sullivan, Ryan and Santos, Luis S and Dieffendahl, Clemens and Horsch, Caroline and Perez-Vicente, Rodrigo and others},
-  journal = {Advances in Neural Information Processing Systems},
-  volume  = {34},
-  pages   = {15032--15043},
-  year    = {2021}
+@inproceedings{terry2021pettingzoo,
+ author      = {Terry, J and Black, Benjamin and Grammel, Nathaniel and Jayakumar, Mario  and Hari, Ananth  and Sullivan, Ryan and Santos, Luis S and Dieffendahl, Clemens and Horsch, Caroline and Perez-Vicente, Rodrigo and Williams, Niall  and Lokesh, Yashas  and Ravi , Praveen },
+ booktitle   = {Advances in Neural Information Processing Systems},
+ editor      = {M. Ranzato and A. Beygelzimer and Y. Dauphin and P.S. Liang and J. Wortman Vaughan},
+ pages       = {15032--15043},
+ publisher   = {Curran Associates, Inc.},
+ title       = {PettingZoo: Gym for Multi-Agent Reinforcement Learning},
+ url         = {https://proceedings.neurips.cc/paper_files/paper/2021/file/7ed2d3454c5eea71148b11d0c25104ff-Paper.pdf},
+ volume      = {34},
+ year        = {2021}
 }
 
-@misc{Terry_PettingZoo_Gym_for,
+@software{Terry_PettingZoo_Gym_for,
   author  = {Terry, Jordan and Black, Benjamin and Grammel, Nathaniel and Jayakumar, Mario and Hari, Ananth and Sullivan, Ryan and Santos, Luis and Perez, Rodrigo and Horsch, Caroline and Dieffendahl, Clemens and Williams, Niall and Lokesh, Yashas},
   license = {MIT},
   title   = {{PettingZoo: Gym for multi-agent reinforcement learning}},
