### 2024-06-03 3f0aabd
Update paper.md
Fix figure sourcing
