+++ paper.bib
@@ -19,9 +19,9 @@
 }
 
 @inbook{iamsar,
-  title        = {International Aeronautical and Maritime Search and Rescue Manual},
+  title        = {Chapter 5. Search techniques and operations},
+  booktitle    = {International Aeronautical and Maritime Search and Rescue Manual},
   organization = {International Maritime Organization and International Civil Aviation Organization},
-  publisher    = {International Maritime Organization},
   year         = {2022},
   volume       = {II},
   chapter      = {5},
