+++ paper.bib
@@ -77,6 +77,20 @@
   abstract = {The prevalence of maritime transportation and operations is increasing, leading to a gradual increase in drowning accidents at sea. In the context of maritime search and rescue (SAR), it is essential to develop effective search plans to improve the survival probability of persons-in-water (PIWs). However, conventional SAR search plans typically use predetermined patterns to ensure complete coverage of the search area, disregarding the varying probabilities associated with the PIW distribution. To address this issue, this study has proposed a maritime SAR vessel coverage path planning framework (SARCPPF) suitable for multiple PIWs. This framework comprises three modules, namely, drift trajectory prediction, the establishment of a multilevel search area environment model, and coverage search. First, sea area-scale drift trajectory prediction models were employed using the random particle simulation method to forecast drift trajectories. A hierarchical probability environment map model was established to guide the SAR of multiple SAR units. Subsequently, we integrated deep reinforcement learning with a reward function that encompasses multiple variables to guide the navigation behavior of ship agents. We developed a coverage path planning algorithm aimed at maximizing the success rates within a limited timeframe. The experimental results have demonstrated that our model enables vessel agents to prioritize high-probability regions while avoiding repeated coverage.}
 }
 
+@article{SILVER2021103535,
+    title          = {Reward is enough},
+    journal        = {Artificial Intelligence},
+    volume         = {299},
+    pages          = {103535},
+    year           = {2021},
+    issn           = {0004-3702},
+    doi            = {https://doi.org/10.1016/j.artint.2021.103535},
+    url            = {https://www.sciencedirect.com/science/article/pii/S0004370221000862},
+    author         = {David Silver and Satinder Singh and Doina Precup and Richard S. Sutton},
+    keywords       = {Artificial intelligence, Artificial general intelligence, Reinforcement learning, Reward},
+    abstract       = {In this article we hypothesise that intelligence, and its associated abilities, can be understood as subserving the maximisation of reward. Accordingly, reward is enough to drive behaviour that exhibits abilities studied in natural and artificial intelligence, including knowledge, learning, perception, social intelligence, language, generalisation and imitation. This is in contrast to the view that specialised problem formulations are needed for each ability, based on other signals or objectives. Furthermore, we suggest that agents that learn through trial and error experience to maximise reward could learn behaviour that exhibits most if not all of these abilities, and therefore that powerful reinforcement learning agents could constitute a solution to artificial general intelligence.}
+}
+
 @article{gmd-11-1405-2018,
   author  = {Dagestad, K.-F. and R\"ohrs, J. and Breivik, {\O}. and {\AA}dlandsvik, B.},
   title   = {OpenDrift v1.0: a generic framework for trajectory modelling},
