+++ paper.bib
@@ -29,6 +29,17 @@
   url          = {https://store.icao.int/en/international-aeronautical-and-maritime-search-and-rescue-manual-volume-ii-mission-co-ordination-doc-9731-2}
 }
 
+@article{trummel1986,
+  title={The complexity of the optimal searcher path problem},
+  author={Trummel, KE and Weisinger, JR},
+  journal={Operations Research},
+  volume={34},
+  number={2},
+  pages={324--327},
+  year={1986},
+  publisher={INFORMS}
+}
+
 @inproceedings{terry2021pettingzoo,
  author      = {Terry, J and Black, Benjamin and Grammel, Nathaniel and Jayakumar, Mario  and Hari, Ananth  and Sullivan, Ryan and Santos, Luis S and Dieffendahl, Clemens and Horsch, Caroline and Perez-Vicente, Rodrigo and Williams, Niall  and Lokesh, Yashas  and Ravi , Praveen },
  booktitle   = {Advances in Neural Information Processing Systems},
