+++ paper.bib
@@ -124,6 +124,12 @@
   issn         = {1983-7402}
 }
 
+@unpublished{algorithmsDSSE2024,
+  author  = {Ricardo Ribeiro Rodrigues and Jorás Custódio Campos de Oliveira and Pedro Henrique Britto Aragão Andrade and Renato Laffranchi Falcão},
+  title   = {Algorithms for Drone Swarm Search (DSSE)},
+  url     = {https://github.com/pfeinsper/drone-swarm-search-algorithms}
+}
+
 @article{dqn2015,
   title    = {Human-level control through deep reinforcement learning},
   author   = {Mnih, Volodymyr and Kavukcuoglu, Koray and Silver, David and
