+++ paper.bib
@@ -127,6 +127,7 @@
 @unpublished{algorithmsDSSE2024,
   author  = {Ricardo Ribeiro Rodrigues and Jorás Custódio Campos de Oliveira and Pedro Henrique Britto Aragão Andrade and Renato Laffranchi Falcão},
   title   = {Algorithms for Drone Swarm Search (DSSE)},
+  year    = {2024},
   url     = {https://github.com/pfeinsper/drone-swarm-search-algorithms}
 }
 
