### 2024-06-03 pfeinsper/drone-swarm-search#260 @renatex333
Hi!

Thank you very much for your detailed and constructive feedback. I have addressed all the points you raised and have made the necessary revisions. Below is a summary of the changes made:

Major Points:

- Summary: I have extended the summary to provide more details on the high-level functionality of the package. This includes the environments they can create and the availability for training reinforcement learning policies.
Additionally, regarding your question about whether to answer if the package provides policies for reinforcement learning in the paper, I have added a brief mention in the summary section for completeness, even though it is detailed in the [official documentation](https://pfeinsper.github.io/drone-swarm-search/Documentation/docsAlgorithms.html) under the "Algorithms" section.

- State of the Field: I have expanded the discussion on the limitations of previous approaches to SAR planning and how reinforcement learning addresses these limitations. Additionally, I have provided more information the metrics used to evaluate SAR path planning solutions. Finally, I have briefly discussed other solutions present on literature and compared our package to them in terms of functionality and availability.

- Structure: I have moved the content after the second paragraph of the Statement of Need section to a new section titled "Functionaluty" for better clarity.

- Implemented RL Algorithms: In the final paragraph, I included a reference to our repository, which contains numerous algorithms implemented by our team using the DSSE package. For detailed descriptions of the implemented algorithms, please refer to the [official documentation](https://pfeinsper.github.io/drone-swarm-search/Documentation/docsAlgorithms.html).

Figures:

- Addressed the compilation issue with Figure 1.

References:

- Fixed the link for the "International Aeronautical and Maritime Search and Rescue Manual" reference and added an ISBN.

- Added the missing URL for the Terry, Black, ... (2021) reference.

You can find all the modifications in this [Pull Request](https://github.com/pfeinsper/drone-swarm-search/pull/261). Please review the changes and let me know if there are any further adjustments needed.

Thank you again for your valuable feedback.
