### 2024-06-08 pfeinsper/drone-swarm-search#259 @draabe
Thanks for resolving!
