### 2024-06-10 pfeinsper/drone-swarm-search#260 @draabe
Thank you, this reads like promising improvements to the paper. Could you kindly update the branch from which the paper is compiled so the changes can be seen on the pdf?
