### 2024-06-10 pfeinsper/drone-swarm-search#258 @draabe
Thank you for clarification.

It is correct that @enricofd , @lfcarrete , @leonardodma are not authors of the submission?
