### 2024-06-10 pfeinsper/drone-swarm-search#262 @draabe
JOSS: repository review
Hi @renatex333,

I've checked the repository for its functionality, and here are my points w.r.t to the revision. In general, installation and core functionality works fine on my machine, and I like the overall structure of the package. The documentation also covers the most important points. I'm having some difficulties running the algorithms with the package being split into two repositories. Here's a few things I noticed:

Major
- I do understand the reasoning behind splitting your package into two repositories, but this makes things a bit harder to use/review. For example, I was trying to run your algorithms on the environment, but wasn't sure how to accomplish this from the documentation. The file paths from the "How to run" section seem outdated, so it's taking some time to find the appropriate files. Furthermore, a short demo on how to combine the DSSE search or coverage environment with the algorithms would be very helpful.
- As a consequence from this, it's a bit opaque which functionality is part of which repository, as the JOSS paper "technically" only aims at the DSSE repo. Is everything from the DSSE-algorithms repo wrapped in DSSE? As above, this could be clarified with a short usage example.
- One more point concerning documentation: I was wondering if your package has any other object than the ``env`` object for high- or low-level usage? This is difficult to see as there seems to be no dedicated API reference in your documentation. If everything is limited to the ``env`` object, I feel that the documentation in the ``Search Environment`` and ``Coverage Environment`` are sufficiently explanatory for usage, but maybe I'm missing some functions?

Minor
- Installation works fine, but unless I'm missing something the GDAL dependency requires Microsoft Visual C++ 14.0 or greater for building. This could be mentioned in your documentation, as for some users this means it's not as easy as running ``pip install DSSE``
- I really appreciate your minimalistic interface design and choice of providing thorough ``Quick Start`` code and then building the explanation of functionality around it. The only thing I did not find sufficient information on (when I want to modify it) is the interface of the ``random_policy`` function - what does this one need to do in order to use it beyond the sample you provide?

Could you kindly help adressing these points? Let me know if anything is unclear.
