### 2024-06-10 #6746 @draabe
Hi @kanishkan91,

thanks for checking in. I've completed the first round of review, and believe the package is rather complete and on the right track. The package works fine, the most important parts are in place, documentation is extensive, and the paper is really well written.

There's been some confusion on my end with respect to the overall functionality and documentation as the package is split across two repositories (cf https://github.com/pfeinsper/drone-swarm-search/issues/262).

Some minor improvements on the paper have already been adressed (https://github.com/pfeinsper/drone-swarm-search/issues/260), as well as clarification on the author list (https://github.com/pfeinsper/drone-swarm-search/issues/258).

@renatex333 Please give ma ping once you've adressed these points and the updated paper compiles, and I'm happy to start a second round of review!

Let me know if anything is unclear in the interim.
