### 2024-06-11 pfeinsper/drone-swarm-search#260 @renatex333
Hi, @draabe!

The paper is compiled from both branches "main" and "paper-writing". I believe the last draft of the paper is available at the Draft PDF workflow run with id [55](https://github.com/pfeinsper/drone-swarm-search/actions/runs/9352369017).
