### 2024-06-11 pfeinsper/drone-swarm-search#258 @renatex333
Hi, @draabe!

Yes, that is correct. Although @enricofd, @lfcarrete, @leonardodma and @Manuel-castanares made significant contributions at the beginning of this project, they are no longer active contributors. Currently, the primary contributors driving the project forward are @JorasOliveira, @Pedro2712, @RicardoRibeiroRodrigues, @fbarth, and myself. This core team has worked extensively on the project, leading to our decision to submit it for publication. However, if you consider it appropriate, we are open to including our former colleagues as authors to acknowledge their early contributions.
