### 2024-06-11 cd652e0
Update docsSearch.md
