+++ docs/Documentation/docsCoverage.md
@@ -15,10 +15,12 @@ The **Coverage Environment** is our second training environment, building on the
 ::: warning Warning
 The DSSE project requires Python version 3.10.5 or higher.
 
-The installation of GDAL (requirement for using opendrift) may need the installation of the following packages
+To install [GDAL](https://gdal.org/) (a requirement for using OpenDrift), you may need to install the following packages:
 ```sh
-sudo apt-get install -y libgdal-dev gdal-bin
+sudo apt install -y libgdal-dev gdal-bin
 ```
+
+For Windows, Microsoft Visual C++ 14.0 or greater is required for building.
 :::
 
 ::: tip Tip
