### 2024-06-14 pfeinsper/drone-swarm-search#266 @wob86
JOSS: running tests
Good to see that within DSSE folder there is a set of tests, but there is no documentation that I found that describes how to run them and in what order (Im assuming its a suite of tests?).  Would be useful way to test code once installed and if anyone in the future makes changes
