### 2024-06-14 #6746 @wob86
#https://github.com/pfeinsper/drone-swarm-search/issues/266
