### 2024-06-14 pfeinsper/drone-swarm-search#266 @renatex333
Hi @wob86,

Thank you for pointing! Indeed, we missed including directions regarding how to run the tests in the DSSE folder. I've now added a brief explanation in the README file on how to run them. Could you please take a look and let me know if you think it's sufficient? It's in PR #267

Thank you for your feedback!
