### 2024-06-14 d762312
Update docs/Documentation/docsAlgorithms.md
Co-authored-by: Renato Laffranchi Falcão <66042074+renatex333@users.noreply.github.com>
