### 2024-06-30 pfeinsper/drone-swarm-search#258 @draabe
Hi, @renatex333

thanks for the clarification. I'll leave the authorship decision to you. Just wanted to make sure the list of authors is appropriate and complete, and I'm convinced with your explanation you're able to make a fair decision.
