### 2024-07-01 #6746 @kanishkan91
👋 @renatex333  - Looks like we are getting close! I'm taking a read through the latest manuscript. In the meanwhile, we need to set up the archive for your new release. Thanks a ton to @draabe and @wob86 for your insightful and timely reviews!

So as next steps for you:

 Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Please ensure that the software archive uses the same license as the license you have posted on GitHub.

 Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.

 Check that software archive exists and title and author list match JOSS paper (or purposefully do not).

 Please respond with the 1) version and 2) DOI of the archived version here
