+++ README.md
@@ -3,6 +3,7 @@
 [![PyPI Release 🚀](https://badge.fury.io/py/DSSE.svg)](https://badge.fury.io/py/DSSE)
 [![License: MIT](https://img.shields.io/badge/License-MIT-brightgreen.svg?style=flat)](https://github.com/pfeinsper/drone-swarm-search/blob/main/LICENSE)
 [![PettingZoo version dependency](https://img.shields.io/badge/PettingZoo-v1.22.3-blue)]()
+[![DOI](https://zenodo.org/badge/599323572.svg)](https://zenodo.org/doi/10.5281/zenodo.12659847)
 ![GitHub stars](https://img.shields.io/github/stars/pfeinsper/drone-swarm-search)
 
 # <img src="https://raw.githubusercontent.com/pfeinsper/drone-swarm-search/main/docs/public/pics/drone.svg" alt="DSSE Icon" width="45" height="25"> Drone Swarm Search Environment (DSSE)
@@ -161,13 +162,22 @@ Our test suite is divided into several parts, each serving a specific purpose:
 If you use this package, please consider citing it with this piece of BibTeX:
 
 ```
-@misc{castanares2023dsse,
-    title={DSSE: a drone swarm search environment}, 
-    author={Manuel Castanares and Luis F. S. Carrete and Enrico F. Damiani and Leonardo D. M. de Abreu and José Fernando B. Brancalion and Fabrício J. Barth},
-    year={2024},
-    eprint={2307.06240},
-    archivePrefix={arXiv},
-    primaryClass={cs.LG},
-    doi={https://doi.org/10.48550/arXiv.2307.06240}
+@software{castanares2023dsse,
+  author       = {Ricardo Ribeiro Rodrigues and
+                  Pedro Henrique Andrade and
+                  Renato Laffranchi Falcão and
+                  Jorás Oliveira and
+                  Manuel Castanares and
+                  Luis Filipe Carrete and
+                  Leonardo Malta and
+                  Enrico Francesco Damiani and
+                  José Fernando Basso Brancalion and
+                  Fabrício Jailson Barth},
+  title        = {DSSE: a drone swarm search environment},
+  month        = jul,
+  year         = 2024,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.12659848},
+  url          = {https://doi.org/10.5281/zenodo.12659848}
 }
 ```
