### 2024-07-05 ce3e621
Create CITATION.cff
