### 2024-07-05 #6746 @kanishkan91
@renatex333 - I'm going to go ahead and recommend accept and hand this over to the EiC for next steps. However, please correct the following in the paper in the meantime-

- [ ] Correct missing and invalid DOIs identified above
- [ ] Correct spelling on "Functonality" on line 46
- [ ] On line 29, this should start with "Trummel & Weisinger" outside the parenthesis
