### 2024-07-05 #6746 @kanishkan91
@renatex333  I have recommended this for acceptance now. Please do correct the above mentioned items in the paper as the AEiC starts looking at the paper.  The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @draabe  and @wob86 for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself. I will provide positive reviews for the reviewers on the JOSS portal given the speed and insight of the reviews.
