### 2024-07-05 #6746 @kanishkan91
@renatex333 Actually, I cannot accept till the above DOI is corrected. Specifically, its this one-

https://doi.org/10.1016/j.artint.2021.103535 is INVALID because of 'https://doi.org/' prefix

Could you correct this in the paper and also add the missing DOIs? I can proceed with the acceptance after that. Thanks!
