### 2024-07-06 #6746 @kanishkan91
@renatex333 Thanks again! I have recommended this for acceptance now (again). The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @draabe and @wob86 for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
