### 2024-07-08 #6746 @crvernon
@kanishkan91 and @renatex333 - it looks like the Zenodo archive references version v3, though the version set in this review thread is v0.2.6.  Could you clear this up?
