### 2024-07-08 #6746 @kanishkan91
@crvernon I see. I think we can use the latest and set the version to v3. Can I still change the version?
