### 2024-07-08 #6746 @crvernon
> @renatex333 - please address the following in the paper:
>
> * [ ]  LINE 43: "Expanding on the state-of-the-art research presented by (Ai et al., 2021; Wu et al., 2024), this..." should read "Expanding on the state-of-the-art research presented by Ai et al. (2021) and Wu et al. (2024), this..." This can be achieved by removing the brackets "[" and "]" from your citation in the paper.  See other JOSS articles if you need help on this.
>
> After the version issue raised above and this citation are cleared up, I will accept this article for publication. Thanks!

Then the one change left should take care of things.  Thanks!
